disable cert verification in http challenge hook
The ACME standard requires that the challenge must be completed via
HTTP, not HTTPS. However, redirects should be followed, and may lead to
an HTTPS url. Let's Encrypt assumes that, when a certificate is first
issued in such a setting, the HTTPS server cannot have a valid
certificate yet, and hence they don't do certificate validation in such
cases. The same rationale is followed here.