send ready signal for all challenges, then verify
This fixes an inefficiency where multiple challenges are prepared and
ready, but the ACME server always only knows about one, before this
script waits for verification. Now, the server is told that all ACME
challenges are ready and can start verifying them. When this script then
starts checking for verification, it should not have to wait as long.